refactor: creating the compositions classes as i saw in examples
