+++ src/beverage/chocolate/hot-chocolate.ts
@@ -1,4 +1,4 @@
-import { Beverage } from "./beverage.interface";
+import { Beverage } from "../beverage.interface";
 
 export class HotChocolate implements Beverage {
 	price(): number {
