+++ src/suplements/cream.ts
@@ -1,0 +1,9 @@
+import { Suplement } from "./suplement.interface";
+
+export class Cream implements Suplement {
+    private PRICE = 0.15;
+
+    getPrice(): number {
+        return this.PRICE;
+    }
+}
