+++ src/suplements/milk.ts
@@ -1,0 +1,10 @@
+import { Suplement } from "./suplement.interface";
+
+export class Milk implements Suplement {
+    private PRICE = 0.1;
+
+    getPrice(): number {
+        return this.PRICE;
+    }
+    
+}
