+++ src/suplements/suplement.interface.ts
@@ -1,0 +1,4 @@
+
+export interface Suplement {
+    getPrice(): number;
+}
