+++ test/hot-chocolate.spec.ts
@@ -1,4 +1,4 @@
-import { HotChocolate } from "../src/beverage/hot-chocolate";
+import { HotChocolate } from "../src/beverage/chocolate/hot-chocolate";
 
 describe("Hot Chocolate cases", () => {
 	it("price should be 1.45", () => {
